gql/requirements.txt
uvicorn[standard]==0.20.0
gunicorn==22.0.0
fastapi[all]==0.104.0
strawberry-graphql==0.235.1
